requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
fastapi==0.86.0
greenlet==2.0.0.post0
h11==0.14.0
idna==3.4
packaging==21.3
psycopg2-binary==2.9.5
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
python-slugify==6.1.2
pyusb==1.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
text-unidecode==1.3
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wrapt==1.14.1
